Correct work with ptrace and external sigs
It's the most interesting note in ptrace' man:
"While  being traced, the child will stop each time a signal is delivered, even if the signal is being ignored."

So, we have to handle signals and (sometimes) reflect 'em to traced process. Now it's OK to break process with ^C or kill -9, for instance.
For signals wich cause stop (SIGSTOP from other, SIGTSTP, SIGTTIN, SIGTTOU) - we break the loop and detach from process.